package/mtd: provide the getrandom() library call if libc doesn't
In order to allow ubihealthd to be built even if libc doesn't provide
getrandom(), we provide our own syscall wrapper if necessary. We run
into this situation with stbgcc-6.3 on ARM and ARM64, since that comes
with glibc 2.24 and getrandom() hasn't been available until glibc 2.25.

While sys_getrandom(), the syscall corresponding to the getrandom()
library function, has been available in Linux since version 3.17 and
the oldest kernel we are using with Buildroot is 4.1, we still have to
account for the case of the syscall itself being undefined. This is due
to the fact that our MIPS toolchain is using 3.12 kernel headers, and
sys_getrandom() is missing in those.

Theoretically, it would work to define the syscall number in the
mtd-utils patch, alongside the library function. The kernel does know
the syscall, after all. However, that would lead to further
complications, as this would need to be architecture specific. Since
ubihealthd doesn't work on kernels older than 5.1 anyway, there is no
need to go down that route. For simplicity, we simply have getrandom()
return ENOSYS if the kernel headers don't define the syscall number for
getrandom() -- even if the kernel could handle the actual request.

refs #SWLINUX-6033

Signed-off-by: Markus Mayer <mmayer@broadcom.com>