busybox: look for older BRCMSTB patches to apply
In order to avoid creating directories for every sub-release of busybox
(increments of the third version number), we have the build system look
for all patch directories that differ only in the last number, starting
from the highest and proceeding to ".0" if necessary.

For a hypothetical Busybox release 1.33.2, we would therefore check for
    * brcmstb-1.33.2
    * brcmstb-1.33.1
    * brcmstb-1.33.0
in this order and try to apply the first set of patches we find.

Not only does this keep the source tree cleaner, it also makes
maintaining the tree easier, as one doesn't need to remember to create
a patch directory for every sub-release of busybox.

refs #SWLINUX-6048

Signed-off-by: Markus Mayer <mmayer@broadcom.com>