dropbear: Allow disabling the use of getrandom()
getrandom() is designed to block until sufficient entropy is collected.
Unfortunately our devices are pathologically equipped with little
entropy on boot. This typically results in the Linux's crng not being
initialized until 2-3 minutes after boot and dropbear not being able to
finish establishing connections. This is unusable.

Make it possible to disable getrandom() by introducing a configuration
option.

refs #SWLINUX-6065

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>