br_config.pl: improve the method for detecting the C library
We used to rely on the compiler name containing the C library flavour.
We also relied on using a generic compiler name that is a symlink to
the actual compiler binary.

For example:
    aarch64-linux-gcc -> aarch64-unknown-linux-musleabi-gcc

This is rather fragile, since aarch64-linux-gcc may not be a symlink
(although it should be), and the actual compiler binary may not contain
the name of the C library (although it should). This approach also
necessitates using the more generic compiler name for the libc
detection to work. So, using CC=aarch64-linux-gcc would work whereas
using CC=aarch64-unknown-linux-musleabi-gcc would fail, which the user
would certainly not expect.

While the aforementioned assumptions will likely be met in most
situations and most users would not encounter any of the issues
described, it is also very easy to remove these restrictions
altogether. Instead of looking at the compiler binary's name, we can
call "$CC -v" and parse the output to find the exact same information
without relying on symlinks and compilers being named and called a
certain way.

refs #SWLINUX-6089

Signed-off-by: Markus Mayer <mmayer@broadcom.com>