misc/Makefile.brcmstb: use git archive to generate source tar-balls
Using "git archive" allows us to use less complex commands to generate
or source tar-balls. It is also more resilient against picking up stray
files. By design, "git archive" won't include anything in the tar-ball
that isn't in the repository.

We retain the usage of "tar" to generate the source tar-balls as a
fallback for situations when a build is performed on a source tree that
isn't checked into a GIT repository.

refs #SWLINUX-6096

Signed-off-by: Markus Mayer <mmayer@broadcom.com>