misc/Makefile.brcmstb: rename BZIP2 make variable
Rename the make variable BZIP2 to BRCM_COMPRESS. The reason for this is
that bzip2 looks for environment variables called BZIP and BZIP2. It'll
take the contents from both and treat them equivalent to command line
arguments.

Since "make" exports its variables via the environment to programs it
calls, bzip2 will see make's BZIP2 variable. Since make's BZIP2 points
to the bzip2 binary, bzip2 will think that it is supposed to compress
itself rather than the data stream on stdin. Of course, this is not the
intention. It also doesn't work and causes the build to fail.

To avoid this clash in variable naming, let's rename the make variable
to something bzip2 doesn't care about.

refs #SWLINUX-6177

Signed-off-by: Markus Mayer <mmayer@broadcom.com>