bin/apply_brcm_patches.sh: add script
Move the application logic for BRCM patches, previously implemented in
busybox.mk, into its own script.

The benefit is twofold. Firstly, the shell script was already on the
complex side for being implemented inside a makefile. Secondly, by
creating a standalone script, it is easy to reuse this logic in other
packages without needing to duplicate any code.

refs #SWLINUX-6230

Signed-off-by: Markus Mayer <mmayer@broadcom.com>