bin/apply_brcm_patches.sh: handle software with two-part version numbers
There are software packages that may use only a two-part version number
(x.y) for some releases. One such software package is util-linux. The
inital release is called x.y rather than x.y.0 (but subsequent releases
will be x.y.1 and so forth).

We need to account for this when we are searching for patches to apply
to a given version.

refs #SWLINUX-6230

Signed-off-by: Markus Mayer <mmayer@broadcom.com>