package/util-linux: provide a new NEWNSID / GETNSID patch for 2.38
Our old patch against util-linux 2.32 no longer applies to util-linux
2.38 due changes in the patch's context. Let's provide a version that
works with 2.38.

Note: The actual contents of the patch has not changed. However, it
has been generated with the help of GIT (which it wasn't before). As a
result, the file name has changed. Also, the patch now contains a
GIT-style commit message.

refs #SWLINUX-6230

Signed-off-by: Markus Mayer <mmayer@broadcom.com>