br_config.pl: get_gcc_dir() handle llvm-wrapper error messages
Since the default expectation going forward is for STBLLVM to be
self-sufficient, llvm-wrapper.pl will always issue an error message
when it is being asked to provide the relevant STBGCC release for the
STBLLVM toolchain in question (as the toolchain won't be relying on
STBGCC).

To spare users unsightly and confusing error messages (that are
expected and don't actually indicate that anything has gone wrong),
let's suppress the error.

refs #SWLINUX-6311

Signed-off-by: Markus Mayer <mmayer@broadcom.com>