misc/Makefile.brcmstb: more resilient way for bundling stbtools tar-ball
Since the introduction of the post-receive hook for the stbtools repo,
tar-balls are being generated automatically and reliably. Therefore,
using these source tar-balls as first choice in releases makes sense.

In the past, generating the tar-ball on the fly was more reliable.
But, with the aforementioned change, this is no longer the case. The
dl/brcm-pl/git repo is no longer being used by default during builds,
so it is also no longer kept up to date by the build system. As a
result, it might be missing relevant commits, and trying to generate
the stbtools tar-ball from it might well fail.

Because it makes sense to use the gen_tar_ball.sh script as a fallback
mechanism, we want to keep it around, but we must ensure the repo is up
to date before we are trying to use it.

Also add some error checking while we are at it.

refs #SWLINUX-6363

Signed-off-by: Markus Mayer <mmayer@broadcom.com>