toolchain/toolchain-wrapper: do pass sysroot to LLVM
We must use Buildroot's copy of the sysroot rather than the original
sysroot bundled with the toolchain. This is because the latter is
read-only while the former is where Buildroot installs libraries and
header files. So, let's remove the special case that would NOT pass
the sysroot from Buildroot to clang.

While we are at it, move -Wno-unused-command-line-argument to the end
of the argument list and add a few comments to better explain the
remaining special cases for LLVM.

refs #SWLINUX-6376

Signed-off-by: Markus Mayer <mmayer@broadcom.com>