package/rsync: force HAVE_C99_VSNPRINTF to "yes"
We know the system will have a modern-enough C library that implements
the required functionality. Since the configure stage can't detect the
system's capabilities (because it is cross-compiling), let's hard-code
the decision, since we know it to be true.

The advantage is that rsync doesn't need to link in its own copy of
snprintf(). Also, using the system implementation of snprintf() may
prevent a compiler warning (or even error) about redefining snprintf()
in rsync.

refs #SWLINUX-6386

Signed-off-by: Markus Mayer <mmayer@broadcom.com>