toolchain/toolchain-external: in 64-bit build copy 32-bit /usr/lib
Check if 32-bit libc.so lives in /usr/lib and, if so, copy the contents
of /usr/lib into the staging area and the rootfs.

Since /usr/bin/libc.so is the actual shared library in the case of
musl-libc, this is needed to support 32-bit binaries in a 64-bit
environment. Conversely, glibc lives in /lib. So, copying /usr/lib is
not needed now with glibc, but this feature could become useful one day
even for glibc, if glibc were to relocate to /usr/lib.

Things are made a little more complicated by the fact that
/usr/lib/libc.so does exist even now in the case of glibc. However, it
is merely a linker file (i.e. ASCII text) pointing to the actual
library in /lib. So, rather than checking whether /usr/lib/libc.so
exists (which is always the case for us), we check whether it is an
ASCII text file to distinguish the two scenarios.

refs #SWLINUX-6398

Signed-off-by: Markus Mayer <mmayer@broadcom.com>