toolchain/toolchain-external: set up ld-musl-armhf.so in 64-bit rootfs
Create the proper symlinks, so /lib/ld-musl-armhf.so.1 is available in
a 64-bit root file system: /lib/ld-musl-armhf.so.1 -> /usr/lib32/libc.so

The original symlink for /lib/ld-musl-armhf.so.1 in the 32-bit sysroot
points to /usr/lib. In a 64-bit environment, the 64-bit libraries live
there, so the original symlink won't work.

This change allows the kernel to invoke /lib/ld-musl-armhf.so.1 upon
encountering a 32-bit binary, and that will call the proper shared
library loader.

refs #SWLINUX-6398

Signed-off-by: Markus Mayer <mmayer@broadcom.com>