board: brcmstb: provide ldd support for 32-bit musl in 64-bit rootfs
musl-libc does not provide an "ldd" command, but implements the feature
within libc.so itself. The ldd command is then provided by creating a
symlink /usr/bin/ldd -> /usr/lib/libc.so.

However, this approach does not work if we have two libc.so (one for
64-bit and one for 32-bit), since there can only be one /usr/bin/ldd.
The two libc variants can't read binaries of the respective other
format. To resolve this issue, we are providing a small wrapper script
that detects word size of the binary and calls the correct libc.so in
ldd mode.

We use od(1) to read the ELF header, since that utility is already
present in the rootfs. Neither file(1) nor readelf(1) are, and adding
them for this purpose alone seems overkill.

refs #SWLINUX-6398

Signed-off-by: Markus Mayer <mmayer@broadcom.com>