package/mmc-utils/Config.in: partially revert c0394cc45446
Upstream commit c0394cc454460ebc99202faa4c10365998e72925 bumps the
mmc-utils version, but it also restricts mmc-utils support to kernels
4.4 and later due to a dependency on some features only present in newer
versions of ioctl.h (which comes with the kernel sources).

We need support for kernel 4.1, since that is the kernel version used by
our toolchains. Doing so is quite straight forward. We are already
providing mmc-utils with its own copy of ioctl.h, so we can support
mmc-utils on 4.1 without issue. We simply revert the portion of commit
c0394cc45446 that restricts mmc-utils support to kernel 4.4 and later.

Note: The kernel header dependency is against the toolchain (and the
headers bundled with it) and not against the actual kernel being built
with the root file system! Due to our toolchains, this means *all* our
builds would be affected by restricting mmc-utils to newer kernels, not
just those using Linux 4.1

refs #SWLINUX-6462

Signed-off-by: Markus Mayer <mmayer@broadcom.com>